# api.yaml
openapi: 3.0.3
info:
  title: Jatayu — Trekking Management Application API
  description: REST API for Jatayu TMA V2 (IIT Madras MAD-II Project)
  version: 2.0.0
servers:
  - url: http://localhost:5000/api
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Trek:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        location: { type: string }
        difficulty: { type: string, enum: [Easy, Moderate, Hard] }
        duration_days: { type: integer }
        total_slots: { type: integer }
        available_slots: { type: integer }
        start_date: { type: string, format: date }
        end_date: { type: string, format: date }
        status: { type: string, enum: [Pending, Approved, Open, Closed, Completed] }
        category: { type: string }
        state: { type: string }
        max_altitude_m: { type: integer }
        permit_required: { type: boolean }
        price_inr: { type: number }
        avg_rating: { type: number }
...
paths:
...
  /treks/:
    get:
      tags: [Public]
      summary: List all open treks
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Search by name or location
        - name: difficulty
          in: query
          schema: { type: string, enum: [Easy, Moderate, Hard] }
        - name: category
          in: query
          schema: { type: string }
      responses:
        '200': { description: List of treks }
  /treks/{id}:
    get:
      tags: [Public]
      summary: Get full trek detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Trek detail with restaurants, agencies, local food }
        '404': { description: Trek not found }

  # ── Admin ─────────────────────────────────────────────────────────
...
  /admin/treks:
    get:
      tags: [Admin]
      summary: List all treks with pagination
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Paginated trek list }
    post:
      tags: [Admin]
      summary: Create a new trek
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Trek' }
      responses:
        '201': { description: Trek created }
  /admin/treks/{id}:
    put:
      tags: [Admin]
      summary: Update a trek
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Trek updated }
    delete:
      tags: [Admin]
      summary: Delete a trek
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Trek deleted }
...
  /staff/treks:
    get:
      tags: [Staff]
      summary: Get treks assigned to this staff member
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: List of assigned treks }
  /staff/treks/{id}/status:
    put:
      tags: [Staff]
      summary: Update trek status
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [Open, Closed, Completed] }
      responses:
        '200': { description: Status updated }
...
  /user/dashboard:
    get:
      tags: [User]
      summary: Trekker dashboard with active bookings
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Dashboard data }
...
  /user/history:
    get:
      tags: [User]
      summary: Get full booking history
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: List of all bookings }
  /user/export-history:
    post:
      tags: [User]
      summary: Trigger CSV export of booking history (background job)
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Export queued, email will be sent }
  /user/profile:
    get:
      tags: [User]
      summary: Get trekker profile
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Profile data }
    put:
      tags: [User]
      summary: Update trekker profile
      security: [{ BearerAuth: [] }]
      responses:
        '200': { description: Profile updated }

  # ── Partners ──────────────────────────────────────────────────────
...